+++ lib/seasonable.rb
@@ -1,8 +1,6 @@
-require 'pry'
 module Seasonable
 
-  # BB (Complete)
-  def biggest_bust(season)
+  def biggest_bust(season) # BB
     teams_reg_season_win_percentage = Hash.new(0)
     teams_post_season_win_percentage = Hash.new(0)
     teams_differences = Hash.new(0)
@@ -38,8 +36,7 @@ module Seasonable
 
   end
 
-  # BB (Complete)
-  def biggest_surprise(season)
+  def biggest_surprise(season) # BB
     teams_reg_season_win_percentage = Hash.new(0)
     teams_post_season_win_percentage = Hash.new(0)
     teams_differences = Hash.new(0)
@@ -75,8 +72,7 @@ module Seasonable
     team_name_finder_helper(team_with_lowest_diff[0])
   end
 
-  # JP (complete)
-  def winningest_coach(season)
+  def winningest_coach(season) # JP
     coach_win_percentage_hash = coach_win_percentage_helper(season)
     best_win_percentage = 0.0
     best_coach = ""
@@ -90,8 +86,7 @@ module Seasonable
     best_coach
   end
 
-  # JP (complete)
-  def worst_coach(season)
+  def worst_coach(season)   # JP
     coach_win_percentage_hash = coach_win_percentage_helper(season)
     worst_win_percentage = 2.0
     worst_coach = ""
@@ -105,26 +100,27 @@ module Seasonable
     worst_coach
   end
 
-  # AM
-  def most_accurate_team(season)
+  def most_accurate_team(season)   # AM
     agg_data = Hash.new(0)
       self.teams.each_pair do |team_id, _|
         agg_data[team_id] = all_shots_season(team_id, season)
       end
-    team_name_finder_helper(agg_data.max_by {|_, v| v}[0])
+
+      agg_data.delete_if do |_, v|
+        v == 0
+      end
+    team_name_finder_helper(agg_data.min_by {|_, v| v if v > 0}[0])
   end
 
-  # AM
-  def least_accurate_team(season)
+  def least_accurate_team(season)   # AM
         agg_data = Hash.new(0)
           self.teams.each_pair do |team_id, _|
             agg_data[team_id] = all_shots_season(team_id, season)
           end
-        team_name_finder_helper(agg_data.min_by {|_, v| v}[0])
+        team_name_finder_helper(agg_data.max_by {|_, v| v}[0])
   end
 
-  # JP
-  def most_tackles(season)
+  def most_tackles(season) # JP
     total_tackles = tackles_helper(season)
     most_tackles = 0
     best_team = 0
@@ -138,8 +134,7 @@ module Seasonable
     team_name_finder_helper(best_team.to_s)
   end
 
-  # JP
-  def fewest_tackles(season)
+  def fewest_tackles(season)   # JP
     total_tackles = tackles_helper(season)
     least_tackles = 10000
     worst_team = 0
@@ -151,82 +146,4 @@ module Seasonable
     end
     team_name_finder_helper(worst_team.to_s)
   end
-
-  ### Helper Methods ###
-
-  def coach_win_percentage_helper(season) #ALL Coaches. Hash. Key = coach name, Value = win percentage
-    coach_array = coach_array_helper
-    coach_win_game_hash = Hash.new(0)
-    coach_win_percentage_hash = Hash.new(0)
-    until coach_array == []
-      coach_win_game_hash[coach_array.shift] = { :wins => 0,
-                                                :games => 0}
-    end
-
-    self.game_teams.each do |game_obj|
-      coach_win_game_hash.each do |coach, win_game_hash|
-        if coach == game_obj.head_coach && season_converter(season) == game_obj.game_id.to_s[0..3].to_i
-          if game_obj.result == "WIN"
-            win_game_hash[:wins] += 1
-            win_game_hash[:games] += 1
-          elsif game_obj.result == "LOSS" || game_obj.result == "TIE"
-            win_game_hash[:games] += 1
-          end
-        end
-      end
-    end
-
-    win_percentage = nil
-    coach_win_game_hash.each do |coach, win_games|
-      win_percentage = ((win_games[:wins]).to_f / (win_games[:games]).to_f).round(2)
-      coach_win_percentage_hash[coach] = win_percentage
-    end
-
-    coach_win_percentage_hash.delete_if do |coach, win_percentage|
-      win_percentage.nan?
-    end
-
-    coach_win_percentage_hash
-  end
-
-  def coach_array_helper #All uniq coaches in an array
-    coach_array = []
-    self.game_teams.each do |game_obj|
-      coach_array << game_obj.head_coach
-    end
-    coach_array.uniq!.sort!
-  end
-
-  def season_converter(season)
-    #convert full season to first 4 characters
-    shortened_season = season[0..3]
-    shortened_season.to_i
-  end
-
-  def all_shots_season(team_id, season)
-
-    all_shots = 0
-    all_goals = 0
-      self.games.each_value do |game|
-        if (game.season == season) && ((game.home_team_id == team_id) || (game.away_team_id == team_id))
-            row = self.game_teams.select do |game_team|
-              ((game_team.game_id == game.game_id) &&
-              (game_team.team_id.to_s == team_id))
-            end
-            if row[0] != nil
-              all_shots += row[0].shots
-              all_goals += row[0].goals
-            end
-        end
-      end
-
-    [all_shots, all_goals]
-
-    if (all_shots > 0) && (all_goals > 0)
-      all_shots.to_f / all_goals
-    else
-      0.00
-    end
-  end
-
 end
